Squashed 'memos/' changes from e4a09c407..17ecfb5cf
17ecfb5cf fix: acl config
c4a24bead chore(frontend): tweak memo relations
d264f4597 fix: code blocks of unknown languages cause HTML injection (#3711)
af952807c feat: write memo UID in file names when exporting to Markdown (#3712)
fefe2b065 chore: fix build artifacts action
5704aaad0 chore: tweak embeded memo style
ef7e2151d chore: update memo relation list
e2fd79200 feat: implement inline memo editor
956f21838 chore: fix popover z-index
878fea5f1 chore: use popover for update tag view setting
5a26b7651 chore: add memo content snippet
0053977b1 chore: bump github.com/aws/aws-sdk-go-v2/config from 1.27.24 to 1.27.26 (#3691)
e5c7571eb chore: bump github.com/aws/aws-sdk-go-v2/feature/s3/manager from 1.17.4 to 1.17.7 (#3690)
daca7193b chore: bump github.com/aws/aws-sdk-go-v2/credentials from 1.17.24 to 1.17.26 (#3693)
ca1d3b6eb chore: bump modernc.org/sqlite from 1.30.1 to 1.30.2 (#3692)
fcdca5b27 chore: bump github.com/aws/aws-sdk-go-v2 from 1.30.1 to 1.30.3 (#3694)
07f79b9f7 chore: bump @mui/joy from 5.0.0-beta.36 to 5.0.0-beta.48 in /web (#3695)
41d9d3bc2 chore: bump prettier from 3.3.2 to 3.3.3 in /web (#3696)
fae951ae2 chore: bump @typescript-eslint/eslint-plugin from 7.15.0 to 7.16.0 in /web (#3697)
446e38f15 fix: memo content snippet
7c9f967a0 chore: bump dompurify from 3.1.5 to 3.1.6 in /web (#3698)
a5d36655e chore: bump tailwindcss from 3.4.4 to 3.4.5 in /web (#3699)
f2cfc528a chore: introduce tag view option
260a4f89f chore: buf generate
9909fd8a6 feat: add snippet field to memo message (#3689)
bcb884324 chore: implement stringify markdown nodes endpoint (#3688)
7c9e54afb chore: update disallow public visibility setting name (#3687)
6825c1def fix: delete access token when sign out
f374169a2 feat: implement html element renderer
4ff3b89a8 chore: update i18n from Weblate (#3680)
5a196e2f1 feat: update gomark version (#3679)
9449886d6 chore: fix linter
457cf92cc feat(api): implement get resource by uid
1ab2c8940 feat(api): implement get memo by uid
34e938a1e chore: bump typescript from 5.5.2 to 5.5.3 in /web (#3664)
b26544506 chore: bump golang.org/x/net from 0.26.0 to 0.27.0 (#3669)
37d13b5cb chore: bump github.com/aws/aws-sdk-go-v2/config from 1.27.23 to 1.27.24 (#3667)
fb4fdb326 chore: bump golang.org/x/crypto from 0.24.0 to 0.25.0 (#3668)
5623a0f06 chore: bump @typescript-eslint/parser from 7.14.1 to 7.16.0 in /web (#3672)
b220f14be chore: bump katex from 0.16.10 to 0.16.11 in /web (#3663)
3e4b33fb2 chore: bump highlight.js from 11.9.0 to 11.10.0 in /web (#3665)
08a43d0fa chore: bump @typescript-eslint/eslint-plugin from 7.14.1 to 7.15.0 in /web (#3666)
36795abac chore: bump golang.org/x/mod from 0.18.0 to 0.19.0 (#3670)
9f7295a4e chore: bump github.com/aws/aws-sdk-go-v2/credentials from 1.17.23 to 1.17.24 (#3671)
626c972d9 chore: add properties to memo detail sidebar
5e3ed201b fix: update mask of update memo (#3659)
eccddb115 chore: add create time to memo detail sidebar (#3657)
a681b8e20 chore: bump github.com/aws/aws-sdk-go-v2/config from 1.27.21 to 1.27.23 (#3640)
881809926 chore: bump github.com/aws/aws-sdk-go-v2/feature/s3/manager from 1.17.1 to 1.17.4 (#3649)
a98d16bca chore: update bug report template fields (#3656)
46ca5bb8e chore: bump github.com/aws/aws-sdk-go-v2/credentials from 1.17.21 to 1.17.23 (#3637)
b1f12d54a chore: bump github.com/aws/aws-sdk-go-v2/service/s3 from 1.56.1 to 1.57.1 (#3644)
4d2b953e2 chore: add enable link preview setting to memo related settings (#3647)
b57b6bae3 chore: bump postcss from 8.4.38 to 8.4.39 in /web (#3642)
90c67d559 chore: bump @typescript-eslint/eslint-plugin from 7.13.1 to 7.14.1 in /web (#3641)
4655c99c4 chore: bump @bufbuild/buf from 1.33.0 to 1.34.0 in /web (#3639)
68e924e48 chore: bump @types/node from 20.14.2 to 20.14.9 in /web (#3638)
ee61bc786 chore: bump react-router-dom from 6.23.1 to 6.24.0 in /web (#3636)
3782d412f chore: upgrade gomark
de754079c chore: bump zustand from 4.5.2 to 4.5.4 in /web (#3632)
df0dcb553 chore: bump typescript from 5.4.5 to 5.5.2 in /web (#3633)
584d73659 chore: bump @reduxjs/toolkit from 2.2.5 to 2.2.6 in /web (#3634)
122710801 chore: bump protobufjs from 7.3.0 to 7.3.2 in /web (#3631)
1ad5d9bf0 chore: update confirm dialog
4a3afffee chore: bump lucide-react from 0.396.0 to 0.399.0 in /web (#3630)
291b81565 feat: implement memo detail sidebar
05c6edfe2 chore: tweak auth pages
6325b3eef chore: bump @types/uuid from 9.0.8 to 10.0.0 in /web (#3616)
80c57692e chore: bump github.com/aws/aws-sdk-go-v2/feature/s3/manager from 1.16.24 to 1.17.1 (#3617)
582863261 chore: bump @typescript-eslint/parser from 7.13.0 to 7.14.1 in /web (#3620)
b923988fd fix: can't change memo UID (#3621)
9b090d909 chore: bump github.com/aws/aws-sdk-go-v2/config from 1.27.18 to 1.27.21 (#3612)
6abc27270 chore: bump lucide-react from 0.395.0 to 0.396.0 in /web (#3614)
3752e7b63 chore: bump eslint-plugin-react from 7.34.2 to 7.34.3 in /web (#3618)
9ba2adb4c chore: bump @typescript-eslint/eslint-plugin from 7.13.0 to 7.13.1 in /web (#3619)
4f4fe9b26 fix: add charset=utf-8 for text content resources (#3608)
ca463891f chore: fix linter warning for slog (#3604)
e33244e64 fix: PATCH requests on /api/v1 not working (#3600)
33cd3ed8e chore: bump version
a09c66986 fix: allow cors for grpc gateway handler (#3574)
759337996 chore: bump lucide-react from 0.383.0 to 0.395.0 in /web (#3584)
cee1d4b1e chore: bump @typescript-eslint/eslint-plugin from 7.12.0 to 7.13.0 in /web (#3585)
77b92f675 chore: bump @bufbuild/buf from 1.32.2 to 1.33.0 in /web (#3586)
795700b16 chore: bump vite from 5.2.13 to 5.3.1 in /web (#3587)
513ff9d2b chore: bump prettier from 3.3.1 to 3.3.2 in /web (#3588)
84692404e chore: bump golang.org/x/oauth2 from 0.20.0 to 0.21.0 (#3583)
d5d0d1472 chore: bump github.com/gorilla/feeds from 1.1.2 to 1.2.0 (#3582)
9bec88eed chore: bump github.com/spf13/cobra from 1.8.0 to 1.8.1 (#3581)
1bf6db815 chore: bump google.golang.org/protobuf from 1.34.1 to 1.34.2 (#3580)
5fb649c3c chore: bump modernc.org/sqlite from 1.30.0 to 1.30.1 (#3579)
b911e98e3 chore: bump docker/build-push-action from 5 to 6 (#3578)
0501f5198 chore: buf format
736637a36 feat: add public flag instead of system setting to enable signup (#3589)
387bf48fc chore: update s3 config
e9d027ca3 chore: bump @typescript-eslint/parser from 7.12.0 to 7.13.0 in /web (#3558)
0c378ec29 chore: bump github.com/aws/aws-sdk-go-v2/feature/s3/manager from 1.16.21 to 1.16.24 (#3547)
5836bf5a7 chore: bump golang.org/x/net from 0.25.0 to 0.26.0 (#3549)
7aba6bef2 chore: bump github.com/aws/aws-sdk-go-v2/config from 1.27.16 to 1.27.18 (#3550)
f31a53caa chore: bump uuid from 9.0.1 to 10.0.0 in /web (#3556)
28624514d chore: bump goreleaser/goreleaser-action from 5 to 6 (#3551)
b42979edc chore: bump golang.org/x/mod from 0.17.0 to 0.18.0 (#3548)
bc1b09581 chore: bump github.com/aws/aws-sdk-go-v2 from 1.27.0 to 1.27.2 (#3546)
2c819ace4 feat: return not found error instead internal error for local file not exist (#3543)

git-subtree-dir: memos
git-subtree-split: 17ecfb5cf5732b2a9033bd84d3796709e08f6dfc

+++ web/src/components/Settings/AccessTokenSection.tsx
@@ -7,13 +7,12 @@ import useCurrentUser from "@/hooks/useCurrentUser";
 import { UserAccessToken } from "@/types/proto/api/v1/user_service";
 import { useTranslate } from "@/utils/i18n";
 import showCreateAccessTokenDialog from "../CreateAccessTokenDialog";
-import { showCommonDialog } from "../Dialog/CommonDialog";
 import Icon from "../Icon";
 import LearnMore from "../LearnMore";
 
 const listAccessTokens = async (name: string) => {
   const { accessTokens } = await userServiceClient.listUserAccessTokens({ name });
-  return accessTokens;
+  return accessTokens.sort((a, b) => (b.issuedAt?.getTime() ?? 0) - (a.issuedAt?.getTime() ?? 0));
 };
 
 const AccessTokenSection = () => {
@@ -38,16 +37,13 @@ const AccessTokenSection = () => {
   };
 
   const handleDeleteAccessToken = async (accessToken: string) => {
-    showCommonDialog({
-      title: "Delete Access Token",
-      content: `Are you sure to delete access token \`${getFormatedAccessToken(accessToken)}\`? You cannot undo this action.`,
-      style: "danger",
-      dialogName: "delete-access-token-dialog",
-      onConfirm: async () => {
-        await userServiceClient.deleteUserAccessToken({ name: currentUser.name, accessToken: accessToken });
-        setUserAccessTokens(userAccessTokens.filter((token) => token.accessToken !== accessToken));
-      },
-    });
+    const confirmed = window.confirm(
+      `Are you sure to delete access token \`${getFormatedAccessToken(accessToken)}\`? You cannot undo this action.`,
+    );
+    if (confirmed) {
+      await userServiceClient.deleteUserAccessToken({ name: currentUser.name, accessToken: accessToken });
+      setUserAccessTokens(userAccessTokens.filter((token) => token.accessToken !== accessToken));
+    }
   };
 
   const getFormatedAccessToken = (accessToken: string) => {
